valid short url

+++ app/controller/url_controller.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from app.database import SessionLocal
 from app.models import UrlModel
 from app.repository.url_repository import url_repository
+from app.utils.hash import valid_hash
 
 
 class UrlRequest(BaseModel):
@@ -36,6 +37,9 @@ def get_db():
 
 @router.get("/{hash}", response_model=UrlResponse)
 def get(hash: str, db: Session = Depends(get_db)):
+    if not valid_hash(hash):
+        raise HTTPException(status_code=400, detail="invalid short url")
+
     record = url_repository.get(hash, db)
     if record is None:
         raise HTTPException(status_code=404, detail="url not found")
@@ -66,6 +70,9 @@ def create(data: UrlRequest, db: Session = Depends(get_db)):
 
 @router.delete("/{hash}", response_model=UrlResponse)
 def delete(hash: str, db: Session = Depends(get_db)):
+    if not valid_hash(hash):
+        raise HTTPException(status_code=400, detail="invalid short url")
+
     record = url_repository.delete(hash, db)
     if record is None:
         raise HTTPException(status_code=404, detail="no register found in db")
@@ -74,6 +81,9 @@ def delete(hash: str, db: Session = Depends(get_db)):
 
 @router.put("/{hash}", response_model=UrlResponse)
 def update(hash: str, data: UrlRequest, db: Session = Depends(get_db)):
+    if not valid_hash(hash):
+        raise HTTPException(status_code=400, detail="invalid short url")
+
     record = url_repository.update(UrlModel(hash=hash, url=str(data.url), on=None), db)
     if record is None:
         raise HTTPException(status_code=404, detail="no register found in db")
@@ -82,6 +92,9 @@ def update(hash: str, data: UrlRequest, db: Session = Depends(get_db)):
 
 @router.put("/activate/{hash}", response_model=UrlResponse)
 def activate(hash: str, db: Session = Depends(get_db)):
+    if not valid_hash(hash):
+        raise HTTPException(status_code=400, detail="invalid short url")
+
     record = url_repository.update(UrlModel(hash=hash, url=None, on=True), db)
     if record is None:
         raise HTTPException(status_code=404, detail="no register found in db")
@@ -90,6 +103,9 @@ def activate(hash: str, db: Session = Depends(get_db)):
 
 @router.put("/deactivate/{hash}", response_model=UrlResponse)
 def deactivate(hash: str, db: Session = Depends(get_db)):
+    if not valid_hash(hash):
+        raise HTTPException(status_code=400, detail="invalid short url")
+
     record = url_repository.update(UrlModel(hash=hash, url=None, on=False), db)
     if record is None:
         raise HTTPException(status_code=404, detail="no register found in db")
